+++ test/function/target/socket.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /*
- * Copyright 2018 Joel E. Anderson
+ * Copyright 2018-2019 Joel E. Anderson
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -165,6 +165,20 @@ namespace {
     target->id = NULL;
 
     stumpless_close_socket_target( target );
+  }
+
+  TEST( SocketTargetCloseTest, Generic ) {
+    struct stumpless_target *target;
+
+    target = stumpless_open_socket_target( "generic-close-test",
+                                           NULL,
+                                           STUMPLESS_OPTION_NONE,
+                                           STUMPLESS_FACILITY_USER );
+    EXPECT_TRUE( target != NULL );
+
+    stumpless_close_target( target );
+
+    EXPECT_TRUE( stumpless_get_error(  ) == NULL );
   }
 
   TEST( SocketTargetCloseTest, NullTarget ) {